Make better double float stuff

+++ include/localization/constants.hpp
@@ -4,8 +4,8 @@ constexpr double MM_TO_IN = 1 / 25.4;
 constexpr double M_TO_IN = 39.3701;
 
 // Values from alexDickhans/echo on GitHub
-constexpr double WALL = 1.78308 * M_TO_IN;
-constexpr double WALL_0_X = WALL;
-constexpr double WALL_1_Y = WALL;
-constexpr double WALL_2_X = -WALL;
-constexpr double WALL_3_Y = -WALL;
+constexpr float WALL = 1.78308 * M_TO_IN;
+constexpr float WALL_0_X = WALL;
+constexpr float WALL_1_Y = WALL;
+constexpr float WALL_2_X = -WALL;
+constexpr float WALL_3_Y = -WALL;
